requirements.txt
django==5.1.4  # Was previously 3.2.25
psycopg2-binary>=2.8
djangorestframework>=3.0
django-rest-knox>=4.0
gunicorn==21.2.0
django-cors-headers
requests
django-debug-toolbar
python-dotenv